refactor #29: 테스트 코드도 패키지 이동
