+++ backend/src/test/java/backend/capstone/domain/mobility/dayroute/service/DayRouteAnalysisLockServiceTest.java
@@ -2,7 +2,7 @@ package backend.capstone.domain.mobility.dayroute.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import backend.capstone.domain.mobility.dayroute.service.DayRouteAnalysisLockService;
+import backend.capstone.domain.mobility.analysis.service.DayRouteAnalysisLockService;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
